requirements.txt
Django>=2.0.4
pytz>=2018.4
psycopg2
Pillow>=5.1.0
djangorestframework>=3.8.2
coreapi>=2.3.3
django-filter>=1.1.0
django-crispy-forms>=1.7.2
django-tinymce
# django-rest-captcha
django-allauth
django-rest-auth
django-simple-captcha
gunicorn
